Dropped support for non desktop apps and added new features
Drop support for Android, macOS, (also a website that was never planned but was in the repository), the reason being that the project is currently mainly focused on desktop and other platforms, see Android and MacOS, this project is not currently running and is not a current priority (perhaps support will be added again in the future).
The addition of the ability to control Klipper printers via the web is not fully functional (only a demo); local control and control via USB will be added. The ability to add, remove, and edit 3D printers in the application will also be added.
This is a summary of the current changes, and for now, the control serves as a "functional demo." (You will be able to control the Klipper printer through the application if you have the URL http://klipper.local or modify it in the code).