Fixed bug where app would be stuck in loop
I changed the logic of updating the UI on the home page due to a bug where changing the status of two printers when starting the application caused it to fail to start and created an infinite loading loop. This was caused by linking the HomePage and PrinterPage ViewModels together and calling functions between them. This was my mistake for rushing development, as I wanted to finish the notifications on the homepage before going to bed. Therefore, I created a separate Service for the logic of calling notifications, I perform all UI manipulation in the given ViewModel, and instead of using the count function for the collection, I use incrementation and decrementation for speed when there are higher numbers of printers. Because of this, I also needed to modify the printer model so that the previous state could be retrieved.

TL;DR: Completely rewritten previous commit, which was rushed and poorly written. It also didn't work, so it needed to be completely rewritten.